graph first version

+++ Library/array.h
@@ -3,7 +3,7 @@
 
 // Non-resizeable array
 // Stores integer values inside
-typedef int Data;
+typedef size_t Data;
 
 struct Array;
